+++ sdk/spring/spring-cloud-azure-service/src/test/java/com/azure/spring/cloud/service/implementation/servicebus/properties/ServiceBusClientCommonTestProperties.java
@@ -19,7 +19,6 @@ public class ServiceBusClientCommonTestProperties extends AzureAmqpSdkProperties
     private String customEndpointAddress;
 
     private boolean shareServiceBusClientBuilder;
-    private Boolean inheritConfiguration;
 
     private String extractFqdnFromConnectionString() {
         if (this.connectionString == null) {
@@ -93,14 +92,5 @@ public class ServiceBusClientCommonTestProperties extends AzureAmqpSdkProperties
 
     public void setShareServiceBusClientBuilder(boolean shareServiceBusClientBuilder) {
         this.shareServiceBusClientBuilder = shareServiceBusClientBuilder;
-    }
-
-    @Override
-    public Boolean getInheritConfiguration() {
-        return inheritConfiguration;
-    }
-
-    public void setInheritConfiguration(Boolean inheritConfiguration) {
-        this.inheritConfiguration = inheritConfiguration;
     }
 }